% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{fe_dimension}
\alias{fe_dimension}
\title{Fixed-effect dimension of a two-way design}
\usage{
fe_dimension(uid, tid, N, T)
}
\arguments{
\item{uid, tid}{integer codes 1..N / 1..T (one per observation)}

\item{N, T}{number of units / periods}
}
\value{
list with `d_K` and `ncomponents`
}
\description{
Union-find on the bipartite unit-time graph:
`d_K = N + T - #connected components`.
}
