% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{print.DesignSummary}
\alias{print.DesignSummary}
\title{Print a design summary}
\usage{
\method{print}{DesignSummary}(x, ...)
}
\arguments{
\item{x}{a \code{DesignSummary}}

\item{...}{unused}
}
\value{
\code{x}, invisibly
}
\description{
Print a design summary
}
