% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_error.R
\name{reliability_from_ratio}
\alias{reliability_from_ratio}
\title{Measurement-error SD implied by an external reliability ratio}
\usage{
reliability_from_ratio(r, within_sd)
}
\arguments{
\item{r}{reliability ratio in (0, 1]}

\item{within_sd}{within-SD of the observed regressor}
}
\value{
scalar measurement-error SD
}
\description{
PSID-style validation studies: `sigma_nu = sqrt((1-r)/r) * within_sd`.
}
\examples{
reliability_from_ratio(0.65, within_sd = 0.15)  # PSID within reliability
}
